Add performance improvements
