+++ x-pack/platform/plugins/shared/alerting/server/application/rule/methods/bulk_create/utils.ts
@@ -1,0 +1,470 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import Boom from '@hapi/boom';
+import Semver from 'semver';
+import { i18n } from '@kbn/i18n';
+import { withSpan } from '@kbn/apm-utils';
+import type { SavedObject, SavedObjectReference } from '@kbn/core/server';
+import type { ActionResult, InMemoryConnector } from '@kbn/actions-plugin/server';
+import type { TaskInstanceWithDeprecatedFields } from '@kbn/task-manager-plugin/server/task';
+
+import { validateAndAuthorizeSystemActions } from '../../../../lib/validate_authorize_system_actions';
+import { RULE_SAVED_OBJECT_TYPE } from '../../../../saved_objects';
+import { parseDuration } from '../../../../../common';
+import { WriteOperations, AlertingAuthorizationEntity } from '../../../../authorization';
+import {
+  validateRuleTypeParams,
+  getRuleNotifyWhenType,
+  getDefaultMonitoringRuleDomainProperties,
+} from '../../../../lib';
+import { getRuleExecutionStatusPending } from '../../../../lib/rule_execution_status';
+import {
+  addGeneratedActionValues,
+  createNewAPIKeySet,
+  extractReferences,
+  validateActions,
+} from '../../../../rules_client/lib';
+import {
+  apiKeyAsAlertAttributes,
+  apiKeyAsRuleDomainProperties,
+} from '../../../../rules_client/common';
+import { ruleAuditEvent, RuleAuditAction } from '../../../../rules_client/common/audit_events';
+import { bulkMarkApiKeysForInvalidation } from '../../../../invalidate_pending_api_keys/bulk_mark_api_keys_for_invalidation';
+import type { RawRule, RuleTypeRegistry, SanitizedRule } from '../../../../types';
+import type { IntervalSchedule } from '../../../../../common';
+import type { BulkOperationError, RulesClientContext } from '../../../../rules_client/types';
+import type { RuleDomain, RuleParams } from '../../types';
+import {
+  transformRuleAttributesToRuleDomain,
+  transformRuleDomainToRule,
+  transformRuleDomainToRuleAttributes,
+} from '../../transforms';
+import { ruleDomainSchema } from '../../schemas';
+import { createRuleDataSchema } from '../create/schemas';
+import type {
+  BulkCreateOperationError,
+  BulkCreateDisabledReason,
+  BulkCreateRulesItem,
+} from './types';
+
+export interface PreparedRule {
+  id: string;
+  name: string;
+  enabled: boolean;
+  rawRule: RawRule;
+  references: SavedObjectReference[];
+  schedule: IntervalSchedule;
+  consumer: string;
+  ruleTypeId: string;
+}
+
+export interface ApiKeyEntry {
+  apiKey: string | null;
+  uiamApiKey: string | null;
+  apiKeyCreatedByUser: boolean | null;
+}
+
+export const getBulkCreateAsDisabledMessage = (message: string): string =>
+  i18n.translate('xpack.alerting.rulesClient.bulkCreate.ruleCreatedDisabledErrorMessage', {
+    defaultMessage: 'Rule created in a disabled state: {message}',
+    values: { message },
+  });
+
+export const collectNewKeysToInvalidate = (entries: Iterable<ApiKeyEntry>): string[] => {
+  const keys: string[] = [];
+  for (const { apiKey, uiamApiKey, apiKeyCreatedByUser } of entries) {
+    if (apiKey && !apiKeyCreatedByUser) keys.push(apiKey);
+    if (uiamApiKey && !apiKeyCreatedByUser) keys.push(uiamApiKey);
+  }
+  return keys;
+};
+
+/**
+ * Pre-fetches every connector referenced by any rule in the batch in a single
+ * `actionsClient.getBulk` call, rather than individually per-rule, to improve performance.
+ */
+export const prefetchActions = async <Params extends RuleParams>({
+  actionsClient,
+  inputsWithIds,
+  logger,
+}: {
+  actionsClient: Awaited<ReturnType<RulesClientContext['getActionsClient']>>;
+  inputsWithIds: Array<{ id: string; rule: BulkCreateRulesItem<Params> }>;
+  logger: RulesClientContext['logger'];
+}): Promise<Map<string, ActionResult | InMemoryConnector> | undefined> => {
+  const unionIds = new Set<string>();
+  for (const { rule } of inputsWithIds) {
+    for (const action of rule.data.actions ?? []) unionIds.add(action.id);
+    for (const action of rule.data.systemActions ?? []) unionIds.add(action.id);
+  }
+
+  if (unionIds.size === 0) return undefined;
+
+  try {
+    const results = await withSpan({ name: 'actionsClient.getBulk', type: 'actions' }, () =>
+      actionsClient.getBulk({ ids: [...unionIds], throwIfSystemAction: false })
+    );
+    return new Map(results.map((r) => [r.id, r]));
+  } catch (err) {
+    // `getBulk` throws synchronously on the first missing id, which would sink the whole batch.
+    // We simply log to console since we can recover later using per-rule connnector fetches.
+    logger.debug(
+      `bulkCreateRules: batched connector pre-fetch failed (${err.message}); falling back to per-rule fetch`
+    );
+    return undefined;
+  }
+};
+
+const sliceActionsById = (
+  actionsById: Map<string, ActionResult | InMemoryConnector> | undefined,
+  actions: ReadonlyArray<{ id: string }>
+): Array<ActionResult | InMemoryConnector> | undefined => {
+  if (!actionsById || actions.length === 0) return undefined;
+  const subset: Array<ActionResult | InMemoryConnector> = [];
+  for (const { id } of actions) {
+    const found = actionsById.get(id);
+    if (found) subset.push(found);
+  }
+  // Returns in the same order as input `actions`
+  return subset;
+};
+
+export const buildTaskInstance = (
+  context: RulesClientContext,
+  prepared: PreparedRule
+): TaskInstanceWithDeprecatedFields => ({
+  id: prepared.id,
+  taskType: `alerting:${prepared.ruleTypeId}`,
+  schedule: prepared.schedule,
+  params: {
+    alertId: prepared.id,
+    spaceId: context.spaceId,
+    consumer: prepared.consumer,
+  },
+  state: {
+    previousStartedAt: null,
+    alertTypeState: {},
+    alertInstances: {},
+  },
+  scope: ['alerting'],
+  // Tasks are scheduled disabled. Phase 5 enables them via taskManager.bulkEnable
+  // so per-task validation drops never produce a running task without a rule SO,
+  // and the activation can randomise schedule datetimes across the batch.
+  enabled: false,
+});
+
+export const toSanitizedRule = <Params extends RuleParams = never>(
+  context: RulesClientContext,
+  so: SavedObject<RawRule>,
+  ruleTypeRegistry: RuleTypeRegistry
+): SanitizedRule<Params> => {
+  const ruleType = ruleTypeRegistry.get(so.attributes.alertTypeId);
+  const ruleDomain: RuleDomain<Params> = transformRuleAttributesToRuleDomain<Params>(
+    so.attributes,
+    {
+      id: so.id,
+      logger: context.logger,
+      ruleType,
+      references: so.references,
+      omitGeneratedValues: false,
+    },
+    context.isSystemAction
+  );
+
+  try {
+    ruleDomainSchema.validate(ruleDomain);
+  } catch (e) {
+    context.logger.warn(`Error validating bulk-created rule domain object for id: ${so.id}, ${e}`);
+  }
+
+  return transformRuleDomainToRule<Params>(ruleDomain, { isPublic: true }) as SanitizedRule<Params>;
+};
+
+interface PrepareRuleArgs<Params extends RuleParams> {
+  context: RulesClientContext;
+  actionsClient: Awaited<ReturnType<RulesClientContext['getActionsClient']>>;
+  actionsById?: Map<string, ActionResult | InMemoryConnector>;
+  username: string | null;
+  id: string;
+  rule: BulkCreateRulesItem<Params>;
+  authzCache: Map<string, Promise<void>>;
+  errors: BulkCreateOperationError[];
+  apiKeysMap: Map<string, ApiKeyEntry>;
+}
+
+export const prepareRule = async <Params extends RuleParams>({
+  context,
+  actionsClient,
+  actionsById,
+  username,
+  id,
+  rule,
+  authzCache,
+  errors,
+  apiKeysMap,
+}: PrepareRuleArgs<Params>): Promise<{ prepared?: PreparedRule; error?: BulkOperationError }> => {
+  const { allowMissingConnectorSecrets } = rule;
+
+  try {
+    const { actions: genActions, systemActions: genSystemActions } = await addGeneratedActionValues(
+      rule.data.actions,
+      rule.data.systemActions,
+      context
+    );
+    const data = { ...rule.data, actions: genActions, systemActions: genSystemActions };
+
+    try {
+      createRuleDataSchema.validate(data);
+    } catch (validationError) {
+      throw Boom.badRequest(`Error validating create data - ${validationError.message}`);
+    }
+
+    // ruleTypeRegistry.get throws 400 if not registered.
+    context.ruleTypeRegistry.get(data.alertTypeId);
+
+    const authzKey = `${data.alertTypeId}::${data.consumer}`;
+    if (!authzCache.has(authzKey)) {
+      authzCache.set(
+        authzKey,
+        context.authorization.ensureAuthorized({
+          ruleTypeId: data.alertTypeId,
+          consumer: data.consumer,
+          operation: WriteOperations.Create,
+          entity: AlertingAuthorizationEntity.Rule,
+        })
+      );
+    }
+    try {
+      await authzCache.get(authzKey)!;
+    } catch (authzError) {
+      context.auditLogger?.log(
+        ruleAuditEvent({
+          action: RuleAuditAction.CREATE,
+          savedObject: { type: RULE_SAVED_OBJECT_TYPE, id, name: data.name },
+          error: authzError,
+        })
+      );
+      throw authzError;
+    }
+
+    context.ruleTypeRegistry.ensureRuleTypeEnabled(data.alertTypeId);
+    const ruleType = context.ruleTypeRegistry.get(data.alertTypeId);
+    const validatedRuleTypeParams = validateRuleTypeParams(data.params, ruleType.validate.params);
+    const actions = sliceActionsById(actionsById, data.actions);
+    const systemActions = sliceActionsById(actionsById, data.systemActions ?? []);
+    const allActionsPrefetched =
+      actionsById !== undefined && (actions || systemActions)
+        ? [...(actions ?? []), ...(systemActions ?? [])]
+        : undefined;
+
+    await validateActions(context, ruleType, data, allowMissingConnectorSecrets, actions);
+    await validateAndAuthorizeSystemActions({
+      actionsClient,
+      actionsAuthorization: context.actionsAuthorization,
+      connectorAdapterRegistry: context.connectorAdapterRegistry,
+      systemActions: data.systemActions ?? [],
+      rule: { consumer: data.consumer, producer: ruleType.producer },
+      preFetchedActions: systemActions,
+    });
+
+    const intervalInMs = parseDuration(data.schedule.interval);
+    if (
+      intervalInMs < context.minimumScheduleIntervalInMs &&
+      context.minimumScheduleInterval.enforce
+    ) {
+      throw Boom.badRequest(
+        `Error creating rule: the interval is less than the allowed minimum interval of ${context.minimumScheduleInterval.value}`
+      );
+    }
+    if (
+      intervalInMs < context.minimumScheduleIntervalInMs &&
+      !context.minimumScheduleInterval.enforce
+    ) {
+      context.logger.warn(
+        `Rule schedule interval (${data.schedule.interval}) for "${ruleType.id}" rule type with ID "${id}" is less than the minimum value (${context.minimumScheduleInterval.value}). Running rules at this interval may impact alerting performance. Set "xpack.alerting.rules.minimumScheduleInterval.enforce" to true to prevent creation of these rules.`
+      );
+    }
+
+    // Mint API key for enabled rules.
+    // Soft-fail: a key-mint failure does NOT reject the rule. We persist it as
+    // disabled, push a degraded error so the caller knows, and let downstream
+    // phases treat it as a never-enabled rule. Re-enabling later will re-mint.
+    let effectiveEnabled = data.enabled;
+    let apiKeyProps:
+      | ReturnType<typeof apiKeyAsRuleDomainProperties>
+      | Awaited<ReturnType<typeof createNewAPIKeySet>> = apiKeyAsRuleDomainProperties(
+      null,
+      username,
+      false
+    );
+    if (data.enabled) {
+      try {
+        apiKeyProps = await createNewAPIKeySet(context, {
+          id: ruleType.id,
+          ruleName: data.name,
+          username,
+          shouldUpdateApiKey: true,
+          errorMessage: 'Error creating rule: could not create API key',
+        });
+        apiKeysMap.set(id, {
+          apiKey: apiKeyProps.apiKey ?? null,
+          uiamApiKey: apiKeyProps.uiamApiKey ?? null,
+          apiKeyCreatedByUser: apiKeyProps.apiKeyCreatedByUser ?? null,
+        });
+      } catch (apiKeyErr) {
+        effectiveEnabled = false;
+        apiKeyProps = apiKeyAsRuleDomainProperties(null, username, false);
+        errors.push({
+          message: getBulkCreateAsDisabledMessage(apiKeyErr.message),
+          status: apiKeyErr.output?.statusCode,
+          rule: { id, name: data.name },
+          disabledReason: 'api_key_creation_failed',
+        });
+      }
+    }
+
+    const allActions = [...data.actions, ...(data.systemActions ?? [])];
+    const artifacts = data.artifacts ?? {};
+    const {
+      references,
+      params: updatedParams,
+      actions: actionsWithRefs,
+      artifacts: artifactsWithRefs,
+    } = await extractReferences(
+      context,
+      ruleType,
+      allActions,
+      validatedRuleTypeParams,
+      artifacts,
+      allActionsPrefetched
+    );
+
+    const createTime = Date.now();
+    const lastRunTimestamp = new Date();
+    const legacyId = Semver.lt(context.kibanaVersion, '8.0.0') ? id : null;
+    const notifyWhen = getRuleNotifyWhenType(data.notifyWhen ?? null, data.throttle ?? null);
+    const throttle = data.throttle ?? null;
+    const { systemActions: _sa, actions: _a, ...restData } = data;
+
+    const ruleAttributes = transformRuleDomainToRuleAttributes({
+      actionsWithRefs,
+      artifactsWithRefs,
+      rule: {
+        ...restData,
+        ...apiKeyProps,
+        enabled: effectiveEnabled,
+        id,
+        createdBy: username,
+        updatedBy: username,
+        createdAt: new Date(createTime),
+        updatedAt: new Date(createTime),
+        snoozeSchedule: [],
+        muteAll: false,
+        mutedInstanceIds: [],
+        notifyWhen,
+        throttle,
+        executionStatus: getRuleExecutionStatusPending(lastRunTimestamp.toISOString()),
+        monitoring: getDefaultMonitoringRuleDomainProperties(lastRunTimestamp.toISOString()),
+        revision: 0,
+        running: false,
+      } as unknown as RuleDomain<Params>,
+      params: { legacyId, paramsWithRefs: updatedParams },
+    });
+
+    if (effectiveEnabled) {
+      ruleAttributes.lastEnabledAt = new Date(createTime).toISOString();
+      ruleAttributes.scheduledTaskId = id;
+    }
+
+    const prepared = {
+      id,
+      name: data.name,
+      enabled: effectiveEnabled,
+      rawRule: ruleAttributes,
+      references,
+      schedule: data.schedule,
+      consumer: data.consumer,
+      ruleTypeId: data.alertTypeId,
+    };
+    return { prepared };
+  } catch (err) {
+    const error = {
+      message: err.message,
+      status: err.output?.statusCode,
+      rule: { id, name: rule.data?.name ?? 'n/a' },
+    };
+    return { error };
+  }
+};
+
+/**
+ * Demote in-memory (enabled -> disabled): flips a set of currently-enabled
+ * prepared rules to disabled, queues their API keys for invalidation, records a degraded
+ * error so the caller can surface "rule was created in a disabled state".
+ */
+export const demotePreparedRules = ({
+  ids,
+  reason,
+  message,
+  preparedRules,
+  apiKeysMap,
+  keysToInvalidate,
+  errors,
+  username,
+}: {
+  ids: string[];
+  reason: BulkCreateDisabledReason;
+  message: string;
+  preparedRules: Map<string, PreparedRule>;
+  apiKeysMap: Map<string, ApiKeyEntry>;
+  keysToInvalidate: Set<string>;
+  errors: BulkCreateOperationError[];
+  username: string | null;
+}): void => {
+  for (const id of ids) {
+    const prepared = preparedRules.get(id);
+    if (!prepared || !prepared.enabled) continue;
+
+    const apiKey = apiKeysMap.get(id);
+    if (apiKey) {
+      for (const k of collectNewKeysToInvalidate([apiKey])) keysToInvalidate.add(k);
+      apiKeysMap.delete(id);
+    }
+
+    // Re-shape `rawRule` to the disabled-rule form.
+    const nullKey = apiKeyAsAlertAttributes(null, username, false);
+    prepared.rawRule = {
+      ...prepared.rawRule,
+      ...nullKey,
+      uiamApiKey: null,
+      enabled: false,
+    };
+    delete prepared.rawRule.scheduledTaskId;
+    delete prepared.rawRule.lastEnabledAt;
+    prepared.enabled = false;
+
+    errors.push({
+      message: getBulkCreateAsDisabledMessage(message),
+      rule: { id, name: prepared.name },
+      disabledReason: reason,
+    });
+  }
+};
+
+export const flushKeysToInvalidate = async (
+  keysToInvalidate: Set<string>,
+  context: RulesClientContext
+): Promise<void> => {
+  if (keysToInvalidate.size === 0) return;
+  await bulkMarkApiKeysForInvalidation(
+    { apiKeys: [...keysToInvalidate] },
+    context.logger,
+    context.unsecuredSavedObjectsClient
+  );
+  keysToInvalidate.clear();
+};
