+++ x-pack/platform/plugins/shared/alerting/server/rules_client/lib/extract_references.ts
@@ -6,6 +6,7 @@
  */
 
 import type { SavedObjectReference } from '@kbn/core/server';
+import type { ActionResult, InMemoryConnector } from '@kbn/actions-plugin/server';
 import type { RuleTypeParams, Artifacts } from '../../types';
 import type { UntypedNormalizedRuleType } from '../../rule_type_registry';
 import type { DenormalizedAction, NormalizedAlertActionWithGeneratedValues } from '../types';
@@ -22,7 +23,8 @@ export async function extractReferences<
   ruleType: UntypedNormalizedRuleType,
   ruleActions: NormalizedAlertActionWithGeneratedValues[],
   ruleParams: Params,
-  ruleArtifacts: Artifacts
+  ruleArtifacts: Artifacts,
+  preFetchedActions?: Array<ActionResult | InMemoryConnector>
 ): Promise<{
   actions: DenormalizedAction[];
   artifacts: Required<DenormalizedArtifacts>;
@@ -32,7 +34,8 @@ export async function extractReferences<
   const actionsClient = await context.getActionsClient();
   const { references: actionReferences, actions } = await denormalizeActions(
     actionsClient,
-    ruleActions
+    ruleActions,
+    preFetchedActions
   );
 
   const { artifacts, references: artifactReferences } = denormalizeArtifacts(ruleArtifacts);
